requirements.txt
apache-airflow==2.8.0
pandas==2.1.4
sqlalchemy==1.4.50
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
great-expectations==0.18.8
pytest==7.4.3